requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile
#
authlib==1.7.0
    # via -r requirements.in
blinker==1.9.0
    # via
    #   -r requirements.in
    #   flask
build==1.4.4
    # via
    #   -r requirements.in
    #   pip-tools
certifi==2026.4.22
    # via
    #   -r requirements.in
    #   requests
cffi==2.0.0
    # via
    #   -r requirements.in
    #   cryptography
charset-normalizer==3.4.7
    # via
    #   -r requirements.in
    #   requests
click==8.3.3
    # via
    #   -r requirements.in
    #   flask
    #   pip-tools
colorama==0.4.6
    # via
    #   -r requirements.in
    #   build
    #   click
cryptography==47.0.0
    # via
    #   -r requirements.in
    #   authlib
    #   joserfc
flask==3.1.3
    # via -r requirements.in
idna==3.13
    # via
    #   -r requirements.in
    #   requests
itsdangerous==2.2.0
    # via
    #   -r requirements.in
    #   flask
jinja2==3.1.6
    # via
    #   -r requirements.in
    #   flask
joserfc==1.6.4
    # via
    #   -r requirements.in
    #   authlib
markupsafe==3.0.3
    # via
    #   -r requirements.in
    #   flask
    #   jinja2
    #   werkzeug
packaging==26.2
    # via
    #   -r requirements.in
    #   build
    #   wheel
pip-tools==7.5.3
    # via -r requirements.in
pycparser==3.0
    # via
    #   -r requirements.in
    #   cffi
pyproject-hooks==1.2.0
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
python-dotenv==1.2.2
    # via -r requirements.in
requests==2.33.1
    # via -r requirements.in
urllib3==2.6.3
    # via
    #   -r requirements.in
    #   requests
werkzeug==3.1.8
    # via
    #   -r requirements.in
    #   flask
wheel==0.47.0
    # via
    #   -r requirements.in
    #   pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
